model/requirements.txt
tensorflow
joblib
numpy
matplotlib
scipy
jupyterlab
tqdm
flax
optax
google-metrax
fastapi
uvicorn
python-multipart
pydantic
